Thread compatible
Making Vineo thread compatible. Currently only OpenGL has problems with threads.The solution was to split vineoDecode() in vineoDecode() and vineoFlush(). In vineoFlush() are OpenGL calls for writing new texture data.

The flushing works in threads. But there are still problems in vineoOpen() used in threads. These need to be solved!