Threading works partially
Not only OpenGL needs a work-a-round for threading support. FFmpeg needs it to. The sollution was to mutex almost every FFmpeg function. Which is lots of work. Thankfully if you compile FFmpeg with threading support (pthreads) and use av_lockmgr_register() function the mutex is done by FFmpeg. Thanks FFmpeg!

Some other problem popped up! Why are the PNG's and JPG's loaded from the internet not working? I didn't test localy. So that need some testing to.