Thread safe!
With the last commits I was already working on thread saving. But now I've got it working and it works very well indeed! In the platform_win32_multithread I created an example to use Vineo with pthreads. I used a number of techniques but this example seems to push the most frames per second. OpenGL didn't like swapping contexes and FFmpeg is thread safe but you cannot run avcodec_decode_video at the same time in multiple threads. It's seems that a simple decode while loop just does the job because phtreads automatically makes some sort of decode queue. Which is awesome! :)